Improve handling of prune threshold changes.
This isn't a complete solution, but it fixes most of the obvious issues.
First, whenever the fetch initial data code runs, it clears the reports.
This ensures we're always starting with a clean slate when importing
historical data. Second, when we change the prune threshold, we
immediately prune instead of waiting for the timer.

A more robust solution will probably need to handle automatically
sorting incoming reports and avoiding duplicates, but it shouldn't
realistically be necessary.

Signed-off-by: Jason Lynch <jason@aexoden.com>